I think its strings that need to be int

+++ src/endpoints/signin.php
@@ -7,14 +7,14 @@ if (isset($_REQUEST['password'])) {
   $userData = getUserData($_REQUEST['password']);
   if($userData[2] == "FALSE"){
     $values = [
-      ["TRUE", time(), userData[4]]
+      ["TRUE", time(), (int) userData[4]]
     ];
     $range = ("C" . getUserRow($_REQUEST['password'])) . (":E" . getUserRow($_REQUEST['password']));
     changeData($values, $range);
     echo $userData[0];
   } else if($userData[2] == "TRUE"){
     $values = [
-      ["FALSE", time(), userData[4] + time() - userData[4]]
+      ["FALSE", time(), (int) userData[4] + time() - (int) userData[4]]
     ];
     $range = ("C" . getUserRow($_REQUEST['password'])) . (":E" . getUserRow($_REQUEST['password']));
     changeData($values, $range);
